move the type to the component and remove unnecessary file
