+++ client/components/stats-date-control/index.tsx
@@ -6,7 +6,20 @@ import qs from 'qs';
 import { useLocalizedMoment } from 'calypso/components/localized-moment';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import DateControl from '../date-control';
-import { StatsDateControlProps } from './types';
+import { DateControlPickerShortcut } from '../date-control/types';
+interface StatsDateControlProps {
+	slug: string;
+	queryParams: string;
+	period: 'day' | 'week' | 'month' | 'year';
+	dateRange: any;
+	shortcutList: DateControlPickerShortcut[];
+	overlay?: JSX.Element;
+	onGatedHandler: (
+		events: { name: string; params?: object }[],
+		event_from: string,
+		stat_type: string
+	) => void;
+}
 
 // Define the event name keys for tracking events
 type EventNameKey =
